c++: BigInteger: Initial commit
A C++ class to handle large integers which cannot
be stored in built-in data types. It is neither optimized
for size, not for speed. It is just written for learning
purpose. The math calculations are emulated the way humans
do basic arithmetic. May contain some bugs, but basic
operations are working.

Signed-off-by: Vishal Thanki <vishalthanki@gmail.com>